+++ src/constants/userTypes.ts
@@ -1,0 +1,12 @@
+const userTypes = [
+  "Admin",
+  "Student",
+  "Teacher",
+  "Accounting",
+  "Assessment",
+  "Dean",
+  "Registrar",
+  "Clinic",
+];
+
+export default userTypes;