+++ src/guards/AuthGuard.tsx
@@ -1,36 +1,83 @@
-import { useState, ReactNode } from "react";
-import { Navigate, useLocation, useNavigate } from "react-router-dom";
-// hooks
-// import useAuth from '../hooks/useAuth';
-// pages
+import { useEffect, ReactNode } from "react";
+import { useMutation } from "react-query";
+import { verify } from "../utils/server/auth";
 import Login from "../pages/Login";
 import Cookies from "js-cookie";
+import Loading from "../components/Loading";
+import { connect } from "react-redux";
+import { setAuthenticatedUser } from "../actions/authenticatedUserActions";
+import { toast } from "react-toastify";
 
 // ----------------------------------------------------------------------
 
 type AuthGuardProps = {
   children: ReactNode;
+  setAuthenticatedUser: Function;
 };
 
-export default function AuthGuard({ children }: AuthGuardProps) {
-  // const { isAuthenticated } = useAuth();
-  const { pathname } = useLocation();
-  const [requestedLocation, setRequestedLocation] = useState<string | null>(
-    null
-  );
+function AuthGuard({ children, setAuthenticatedUser }: AuthGuardProps) {
   const sessionToken = Cookies.get("sessionToken");
+  const {
+    mutate: triggerTokenVerify,
+    isLoading: isTriggerTokenVerifyLoading,
+    isError: isTriggerTokenVerifyError,
+    isSuccess: isTriggerTokenVerifySuccess,
+    data: triggerTokenVerifyData,
+  } = useMutation(async (tokenVerify: any) => verify(tokenVerify));
 
-  if (sessionToken == null) {
-    if (pathname !== requestedLocation) {
-      setRequestedLocation(pathname);
+  useEffect(() => {
+    if (sessionToken) {
+      triggerTokenVerify({ token: sessionToken });
     }
-    return <Login />;
-  }
+  }, [sessionToken, triggerTokenVerify]);
 
-  if (requestedLocation && pathname !== requestedLocation) {
-    setRequestedLocation(null);
-    return <Navigate to={requestedLocation} />;
-  }
+  useEffect(() => {
+    if (isTriggerTokenVerifySuccess) {
+      const { _id, userType } = triggerTokenVerifyData.user;
+      setAuthenticatedUser({
+        id: _id,
+        type: userType,
+        name: triggerTokenVerifyData.name,
+      });
+    }
+  }, [
+    setAuthenticatedUser,
+    triggerTokenVerifyData,
+    isTriggerTokenVerifySuccess,
+  ]);
 
-  return <>{children}</>;
+  const renderComponent = () => {
+    if (sessionToken && isTriggerTokenVerifyLoading) {
+      return <Loading />;
+    } else if (
+      sessionToken &&
+      !isTriggerTokenVerifyLoading &&
+      !isTriggerTokenVerifyError &&
+      isTriggerTokenVerifySuccess
+    ) {
+      return children;
+    } else {
+      if (
+        sessionToken &&
+        !isTriggerTokenVerifyLoading &&
+        isTriggerTokenVerifyError
+      ) {
+        toast.warning(
+          "Oops! Your session is expired. Please login to continue",
+          { toastId: "authWarning" }
+        );
+      } else if (!sessionToken) {
+        toast.warning("Oops! Please login to access that page", {
+          toastId: "authWarning",
+        });
+      }
+      return <Login />;
+    }
+  };
+
+  return <>{renderComponent()}</>;
 }
+
+const mapStateToProps = () => ({});
+
+export default connect(mapStateToProps, { setAuthenticatedUser })(AuthGuard);