+++ src/pages/Settings.tsx
@@ -1,26 +1,34 @@
 import { Icon } from "@iconify/react";
-import { toast } from "react-toastify";
+import { useNavigate } from "react-router-dom";
+import Breadcrumb from "../components/Breadcrumb";
+import { PATH_MODULE } from "../routes/path";
 
 const Settings = () => {
-  const notify = () => toast("Wow so easy !");
+  const navigate = useNavigate();
+  const links = [
+    {
+      name: "Module",
+      path: "/module",
+    },
+    {
+      name: "Settings",
+      path: "/module/settings",
+    },
+  ];
   return (
     <div>
-      <button onClick={notify}>Notifyasdasdasdasd</button>
-      <div className="w-full pl-16 py-10 text-ash-gray">
-        {" "}
-        Module{" "}
-        <span>
-          <a className="text-primary">Settings</a>
-        </span>
-      </div>
+      <Breadcrumb links={links} />
       <div className="pl-72 text-ash-gray"> Settings </div>
       <div className="grid grid-cols-6 mx-72 py-8 justify-items-center ;\">
-        <a href="#" className={` hover:bg-gray px-9 py-5 `}>
+        <span
+          onClick={() => navigate(PATH_MODULE.settings.studentType.root)}
+          className={`hover:bg-gray px-9 py-5 hover:cursor-pointer`}
+        >
           <div className="text-center grid justify-items-center">
             {<Icon icon="bi:person-plus" fontSize={45} color="#006fc9" />}
             <span className="text-ash-gray pt-3">Student Type</span>
           </div>
-        </a>
+        </span>
         <a href="#" className={` hover:bg-gray px-9 py-5 `}>
           <div className="text-center grid justify-items-center">
             {<Icon icon="bi:person-video3" fontSize={45} color="#006fc9" />}